When loading a partially-annotated dataset, added the option to skip examples with no labels.
When trainin on two datasets, changed the number of iterations that
comprise an epoch. Before, this was the total number of examples
in both datasets. Now, it is only the number of examples in the
first dataset (partially-labeled, in general). This allows a easier
way to compare results of models training only on completely annotated
and trained on both annotated and partially-annotated datasets.